specs: record requests with VCR
Our tests now replay a live request and its response rather than
mocked data that could be mistaken (and in fact this change proved
that some of our mocked data was faulty).

We might want to re-record cassettes from time to time to make sure
nothing has changed unexpectedly on the LLM side. When recording a
request, the environment variable LLM_SECRET should be set to a valid
value and it will be automatically scrubbed from any cassettes that
are eventually checked into the repository.