ollama: add /embeddings support
I was recently able to setup ollama within my local network, and I
noticed it has support for an /embeddings endpoint similar to OpenAI,
Gemini, etc.

This change implements `LLM::Ollama#embed` on top of the /embeddings
endpoint. So far I was only able to test with the llama3.2 model.