Add stream support via Server-Side Events (SSE)
This change reaches a milestone in llm.rb's journey with support
finally added for streaming. All providers are supported, and streaming
is implemented on top of Server-Side Events (SSE).

We expect that the interface will evolve over time to support yielding
a message to a Proc but for now you can stream to an IO-like object,
or enable streaming without an IO object.