Add xAI (grok) support
Its chat completions API is on the same level as OpenAI and Gemini, and even
surpasses Anthropic in some areas with built-in support for tool calls, JSON
schema generation, vision capabilities, and more. This is shown in its impressive
spec/xai/chat_spec.rb spec coverage.

xAI also provides an OpenAI-style image generation API, putting it in the same
league as OpenAI and Gemini when it comes to image support.

All of this makes xAI a welcome addition to the ecosystem. In testing so far,
I've found its capabilities to be impressive though in some cases, response times
can be a bit slower compared to other providers.