Re-use connections between requests
The LLM::Provider#initialize method can be configured to persist
connections through the `persistent` option. When set to true, the
option configures a net-http client with the [`Net::HTTP::Persistent`](https://github.com/drbrain/net-http-persistent)
gem.

The client is re-used among all instances of LLM::Provider that
share the same host, and the net-http-persistent gem takes care
of keeping the socket alive.

This feature is optional and requires a third party gem to be
installed.