gemini: 'candidates' might be null
When gemini does not generate a candidate, it will not return
an object with a 'candidates' property. This can sometimes happen
if the prompt or image is flagged for safety reasons, but it can
also be a temporary problem.

The fix implements the 'candidates' method in case Gemini do not
implement it for us.

Ref #90