Add LLM::Function.find_by(name:)
A tool that is a subclass of LLM::Tool is stored in LLM.functions
with a key that is equal to the subclass of LLM::Tool (aka an
instance of Class) rather than with a key that is equal to the
tool / function name.

This presents a challenge - how do we look up such functions when
we only have their function name as a string? The solution is a new
method that searches for a function by its name via `Hash#each_value`