Add support for zAI (https://z.ai)
The zAI provider is based in Signapore / China, and it comes across
as especially affordable with a flat fee of 3 USD per month to access
their API and also comes with a free model (glm-4.5-flash). It appears
to be optimized for coding tasks as well. Their models also support
reasoning, and include it separately in the response.